Very big commit with new v2 api with many new features

+++ src/controllers/blogController.ts
@@ -1,39 +1,53 @@
-import { Request, Response } from "express";
-import { Types } from "mongoose";
-import Blog from "../models/Blog.js";
 import bcryptjs from "bcryptjs";
+import { Request, Response } from "express";
+import Blog from "../models/Blog.js";
+import Comment from "../models/Comment.js";
+import Interaction from "../models/Interaction.js";
+// import { blogtypes } from "../../types/types.js";
 const { compare } = bcryptjs;
-import { blogtypes } from "../@types/types.js";
 
 import {
-	BodyRestrictedError,
-	AnonymousDraftError,
-	BlogDoesntExistError,
-	NotBlogAuthorError,
-	getNextUniqueId,
-	restrictedBodyCheck,
-	anonymousDraftCheck,
-	blogExistAndAuthorCheck,
+	buildResponse,
+	generateSlug,
 	getAnonymousUserId,
 	getDeleteIds,
-	generateSlug,
-	buildResponse
+	getNextUniqueId
 } from "../utils/blogUtils.js";
+
+enum blogtypes {
+	NORMAL_DRAFT = 1,
+	ANON_POST,
+	NORMAL_POST
+}
 
 const saveBlog = async (req: Request, res: Response): Promise<void> => {
 	try {
 		// take all parameters from extended request
-		let userId = req.user.id;
-		const { state, anon } = req.validatedData;
-		const { title, content, tags, category } = req.validatedBody;
-		const blogType = state === "draft" ? blogtypes.NORMAL_DRAFT : anon === "true" ? blogtypes.ANON_POST : blogtypes.NORMAL_POST;
+		console.log(res.locals);
+		let userId = res.locals.user.id;
+		const { state, anon } = res.locals.validatedData;
+		const { title, content, tags, category } = res.locals.validatedBody;
+		const blogType = state === "draft" ?
+			blogtypes.NORMAL_DRAFT : anon === "true" ?
+				blogtypes.ANON_POST : blogtypes.NORMAL_POST;
 
 		// some checks
-		restrictedBodyCheck(title, content);
-		anonymousDraftCheck(state, anon);
-
+		const restrictedTitle = "[Deleted Blog]";
+		const restrictedContent = "This blog has been deleted"
+		if (title === restrictedTitle || content === restrictedContent) {
+			res.status(400).json({ message: "This title or content is restricted" });
+			return;
+		}
+		if (state === "draft" && anon === "true") {
+			res.status(400).json({ message: "Anonymous drafts are not allowed" });
+			return;
+		}
 		// generate identifiers
 		const uniqueId = await getNextUniqueId();
+		if (uniqueId === -1) {
+			res.status(400).json({ message: "Internal Server Error" });
+			return;
+		}
 		const slug = generateSlug(title);
 
 		// generate deletion id's if anonymous published post
@@ -45,7 +59,7 @@ const saveBlog = async (req: Request, res: Response): Promise<void> => {
 		}
 
 		// save the new blog
-		const newBlog = new Blog({
+		const newBlog = await Blog.create({
 			author: userId,
 			state,
 			uniqueId,
@@ -56,17 +70,20 @@ const saveBlog = async (req: Request, res: Response): Promise<void> => {
 			category,
 			anonBlogDeleteId: hashedAnonBlogDeleteId
 		});
-		await newBlog.save();
+		// if (blogType !== blogtypes.NORMAL_DRAFT) {
+		// 	await Interaction.create({
+		// 		blogId: newBlog._id,
+		// 		likes: [],
+		// 		dislikes: [],
+		// 		stars: []
+		// 	});
+		// }
 
 		const response = buildResponse(blogType, anonBlogDeleteId);
 		res.status(200).json(response);
 		return;
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
-		if (err instanceof BodyRestrictedError || err instanceof AnonymousDraftError) {
-			res.status(400).json({ message: err.message });
-			return;
-		}
 		res.status(500).json({ message: "Internal Server Error" });
 		return;
 	}
@@ -75,13 +92,20 @@ const saveBlog = async (req: Request, res: Response): Promise<void> => {
 const deleteBlog = async (req: Request, res: Response): Promise<void> => {
 	try {
 		// take all parameters from req
-		const userId = req.user.id;
-		const { state, blogId } = req.validatedData;
+		const userId = res.locals.user.id;
+		const { state, blogId } = res.locals.validatedData;
 
 		// check if blog exists to delete
 		// check if user is author of blog
-		await blogExistAndAuthorCheck(blogId, userId);
-
+		const blog = await Blog.findOne({ _id: blogId });
+		if (!blog) {
+			res.status(404).json({ message: "This blog dosn't exist" });
+			return;
+		}
+		if (!blog.author.equals(userId)) {
+			res.status(403).json({ message: "You are not allowed to delete this blog" });
+			return;
+		}
 		// delete the blog from database if it's a draft
 		if (state === "draft") {
 			await Blog.deleteOne({ _id: blogId });
@@ -91,7 +115,8 @@ const deleteBlog = async (req: Request, res: Response): Promise<void> => {
 				{
 					$set: {
 						title: "[Deleted Blog]",                    // change the title and content
-						content: "This blog has been deleted"
+						content: "This blog has been deleted by user",
+						slug: "deleted-blog"
 					}
 				});
 		}
@@ -99,10 +124,6 @@ const deleteBlog = async (req: Request, res: Response): Promise<void> => {
 		return;
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
-		if (err instanceof BlogDoesntExistError || err instanceof NotBlogAuthorError) {
-			res.status(400).json({ message: err.message });
-			return;
-		}
 		res.status(500).json({ message: "Internal Server Error" });
 		return;
 	}
@@ -112,12 +133,20 @@ const deleteBlog = async (req: Request, res: Response): Promise<void> => {
 
 const deleteAnonBlog = async (req: Request, res: Response): Promise<void> => {
 	try {
-		const { anonBlogdeleteId, blogId } = req.validatedData;
+		const { anonBlogdeleteId, blogId } = res.locals.validatedData;
 		const anonymous_user = getAnonymousUserId();
-		const blog = await blogExistAndAuthorCheck(blogId, anonymous_user);
+		const blog = await Blog.findOne({ _id: blogId });
+		if (!blog) {
+			res.status(404).json({ message: "This blog dosn't exist" });
+			return;
+		}
+		if (!blog.author.equals(anonymous_user)) {
+			res.status(403).json({ message: "You are not authorized to delete this blog" });
+			return;
+		}
 		const isDeleteIdSame = await compare(anonBlogdeleteId, blog.anonBlogDeleteId);
 		if (!isDeleteIdSame) {
-			res.json(400).json({ message: "Invalid deleteId" });
+			res.status(400).json({ message: "Invalid deleteId" });
 			return;
 		}
 		await Blog.updateOne(
@@ -125,16 +154,14 @@ const deleteAnonBlog = async (req: Request, res: Response): Promise<void> => {
 			{
 				$set: {
 					title: "[Deleted Blog]",                    // change the title and content
-					content: "This blog has been deleted"
+					content: "This blog has been deleted by user",
+					slug: "deleted-blog"
 				}
 			});
 		res.status(200).json({ message: `Anonymous blog deleted successfully` });
+		return;
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
-		if (err instanceof BlogDoesntExistError || err instanceof NotBlogAuthorError) {
-			res.status(400).json({ message: err.message });
-			return;
-		}
 		res.status(500).json({ message: "Internal Server Error" });
 		return;
 	}
@@ -144,28 +171,39 @@ const publishDraft = async (req: Request, res: Response): Promise<void> => {
 	try {
 		// check for anonymous user id's existence in .env
 		const anonymous_user = getAnonymousUserId();
-		const userId = req.user.id;
-		const { anon, blogId } = req.validatedData;
+		const userId = res.locals.user.id;
+		const { anon, blogId } = res.locals.validatedData;
 
 		// check if draft blog exists to publish
 		// check if user is author of blog
-		const blog = await blogExistAndAuthorCheck(blogId, userId);
-
+		const blog = await Blog.findOne({ _id: blogId });
+		if (!blog) {
+			res.status(404).json({ message: "This blog dosn't exist" });
+			return;
+		}
+		if (!blog.author.equals(userId)) {
+			res.status(403).json({ message: "You are not allowed to publish this blog" });
+			return;
+		}
 		// if blog is to be published anonymously set the userId to global anonymous id
 		if (anon === "true") blog.author = anonymous_user
 		// change the blog state to published
 		blog.state = "published"
 		await blog.save();
 
+		// create empty interactions and comments collection
+		// await Interaction.create({
+		// 	blogId: blog._id,
+		// 	likes: [],
+		// 	dislikes: [],
+		// 	stars: []
+		// });
+
 		// send appropriate response message
 		if (anon === "true") res.status(200).json({ message: "Draft published anonymously" });
 		else res.status(200).json({ message: "Draft published" });
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
-		if (err instanceof BlogDoesntExistError || err instanceof NotBlogAuthorError) {
-			res.status(400).json({ message: err.message });
-			return;
-		}
 		res.status(500).json({ message: "Internal Server Error" });
 		return;
 	}
@@ -173,16 +211,28 @@ const publishDraft = async (req: Request, res: Response): Promise<void> => {
 
 const updateBlog = async (req: Request, res: Response): Promise<void> => {
 	try {
-		let userId = req.user.id;
-		const { state, blogId } = req.validatedData;
-		const { title, content, tags, category } = req.validatedBody;
+		let userId = res.locals.user.id;
+		const { state, blogId } = res.locals.validatedData;
+		const { title, content, tags, category } = res.locals.validatedBody;
 
 		// check if draft blog exists to publish
 		// check if user is author of blog
-		const blog = await blogExistAndAuthorCheck(blogId, userId);
-
+		const blog = await Blog.findOne({ _id: blogId });
+		if (!blog) {
+			res.status(404).json({ message: "This blog dosn't exist" });
+			return;
+		}
+		if (!blog.author.equals(userId)) {
+			res.status(403).json({ message: "You are not allowed to update this blog" });
+			return;
+		}
 		// check if title update is forbidden or not
-		restrictedBodyCheck(title, content);
+		const restrictedTitle = "[Deleted Blog]";
+		const restrictedContent = "This blog has been deleted"
+		if (title === restrictedTitle || content === restrictedContent) {
+			res.status(400).json({ message: "This title or content is restricted" });
+			return;
+		}
 
 		// update the blog
 		blog.title = title;
@@ -195,12 +245,6 @@ const updateBlog = async (req: Request, res: Response): Promise<void> => {
 		return;
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
-		if (err instanceof BlogDoesntExistError
-			|| err instanceof NotBlogAuthorError
-			|| err instanceof BodyRestrictedError) {
-			res.status(400).json({ message: err.message });
-			return;
-		}
 		res.status(500).json({ message: "Internal Server Error" });
 		return;
 	}
@@ -209,9 +253,9 @@ const updateBlog = async (req: Request, res: Response): Promise<void> => {
 const getBlogsWithSearch = async (req: Request, res: Response): Promise<void> => {
 	try {
 		// take parameters from req
-		const search = req.body.search ? String(req.body.search) : false;
-		const userId = req.user.id;
-		const state = req.validatedData.state;
+		const search = req.query.search ? String(req.query.search) : false;
+		const userId = res.locals.user.id;
+		const state = res.locals.validatedData.state;
 
 		// build the query accordingly if search exists or not 
 		const query = search ?
@@ -227,7 +271,10 @@ const getBlogsWithSearch = async (req: Request, res: Response): Promise<void> =>
 			: { author: userId, state };
 
 		// execute the query
-		const blogs = await Blog.find(query).select("-__v -anonBlogDeleteId").populate("category");
+		const blogs = await Blog.find(query).select("-__v -anonBlogDeleteId").populate([
+			{ path: "author", select: "name" },
+			{ path: "category", select: "name" },
+		]);
 		res.status(200).json(blogs);
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
@@ -239,12 +286,21 @@ const getBlogsWithSearch = async (req: Request, res: Response): Promise<void> =>
 const getBlogById = async (req: Request, res: Response): Promise<void> => {
 	try {
 		// take all parameters from req
-		const userId = req.user.id;
-		const { blogId, state } = req.validatedData;
+		const userId = res.locals.user.id;
+		const { blogId, state } = res.locals.validatedData;
 
 		// here get the blogs by the user only
-		const blog = await Blog.findOne({ author: userId, _id: blogId, state }).select("-__v -anonBlogDeleteId");
-		res.status(200).json(blog)
+		const blog = await Blog.findOne({ author: userId, _id: blogId, state }).select("-__v -anonBlogDeleteId").populate([
+			{ path: "author", select: "name" },
+			{ path: "category", select: "name" },
+		]);
+		const parentComments = await Comment.find({
+			blogId,
+			parentId: blogId  // treating blogId as parentId for top-level comments
+		})
+		.select("-__v")
+		.populate("userId", "name");
+		res.status(200).json({ blog, parentComments });
 	} catch (err) {
 		console.error("❌ Some error occurred:", err);
 		res.status(500).json({ message: "Internal Server Error" });
@@ -253,11 +309,6 @@ const getBlogById = async (req: Request, res: Response): Promise<void> => {
 }
 
 export {
-	saveBlog,
-	deleteBlog,
-	updateBlog,
-	getBlogsWithSearch,
-	getBlogById,
-	publishDraft
-}
+	deleteAnonBlog, deleteBlog, getBlogById, getBlogsWithSearch, publishDraft, saveBlog, updateBlog
+};
